Delete utils
